Fix sendRequest
sendRequest was not returning a value to the channel if there was
_not_ an error. How did this even work ?!?

Also remove server_test.go because it was wrong, it was testing that
sending a properly formed, but invalid (ie, requested wrong extended
packet) packet would cause an error during send ... which is wrong.
sendRequest will only return an error if the client connection was
closed.