Limited upload-only SFTP for POS.
This is an upload-only version of github.com/pkg/sftp. Only the
following operations are allowed: INIT, OPEN, WRITE, CLOSE, and
REALPATH. An open must open the file for writing.

Three hooks have been added as server options:

* Set the upload path (working directory) as seen by the client. Files
can only be uploaded to this directory.

* Provide a filename validator/mapper function. This function is called
before opening a file. This function is used to determine whether the
file name supplied by the client is valid for uploading and to map the
upload file name to the actual file system path at which the file
should be created.

* Provide an upload notification function. This function is called when
a file is closed. The name of the file is passed as an argument. The
function is called in a separate goroutine.