sftp: do not use filepath, but assume unix style paths
Most of the cases where filepath is used, path should actually be used.
sftp may be run as a server on Windows with *nix clients, or run
as *nix server with windows clients. The path sepecification for now
should be agreed to use a common format, forward slashes.

In the future we may want the server to handle clients with paths
from either paths separaters, but for now assume forward slashes
for all uses.

Updates #8