Added support for additional packet types.
Testing with WinSCP revealed that support for a few more packet types
was required. The following are now supported:

LSTAT: Allowed on the upload path only. Returns the upload path as the
name, dir+0755 as the mode, and the current time as the modify time.

OPENDIR: Allowed on the upload path only. Just calls a hook function.

READDIR:  Calls a hook function to provide the list of files to be
returned.

SETSTAT: No-op.