Fix issue with path 'cleaning' and adds tests
Fix #207

The primary issue of the resolved ticket was path cleaning code returned
the path of '/.' instead of '/'.

We also reviewed and clarified the use of filepath/path and how OS
specific paths are handled in the server code. All paths are converted
to POSIX paths as they come into the server. They are then POSIX paths
for all internal operations.