Remove bad test
TestCloseOutOfOrder was supposed to test handling the case where the
client sends a write packet after the close packet for that file. It did
this poorly and the test is unneeded as what should happen is an error,
and this is what does happen. The RFC specifies that once the close
packet is sent, that's it and so it should be up to the client to
properly finish writing before closing.