fix race condition in fileget/fileput
This fixes the concurrent access/locking race in both fileget and
fileput when accessing the saved readerAt/writerAt. It eliminates the
dedicated functions and moves the access/locking inline where they were
called to make it clearer.

Thanks to Marshall Brekka (@marshallbrekka) for the original bug report
and outline for the fix.