Stop ReadFromWithConcurrency submitting out of order writes
Before this change, ReadFromWithConcurrency would very often submit
writes to the server out of order due to small timing differences when
lots of blocks were queued and ready to send.

This causes more work for the server as seeking about the file isn't
free. Writing chunks of a file out of order can also cause
fragmentation and consequently loss of performance on that file.

This patch causes the write to the channel to be handled synchronously
but the acks are received asynchronously allowing for many write
packets to be in flight as before.