Add SetAttributes as an alternate option.
This would replace the SetStat public function from the last commit.

Instead, this makes it practical to construct a Setstat request.

From some quick testing, this appears to do the job correctly, and it
feels like a lot less of hack.

It still doesn't feel entirely right, but it's closer.